[BugFix]Don't return error when decoding min/max is not supported (#59346)
we didn't support decoding min/max value of some types, but we shouldn't return
the error status to users.
On main branch and branch-3.5, there is no bug, we just deal with this error status
asap to avoid further bug.
On branch-3.4 and branch-3.3, there is bug when the type of predicate is in-filter
or rf-min-max and the type of data is float/double.

Signed-off-by: zombee0 <ewang2027@gmail.com>