[Enhancement] Adding catalog level access manager for external catalogs to allow operations on DB and tables (#59611)
Why I'm doing:
When we have external systems like Hive which are already doing authorization and access control we donot have to rely on Starrocks native access control. Also with Ouath and OIDC we can have users that are not registered with Starrocks and NativeAccessControl will fail with them.

What I'm doing:
Added AllowAllHiveAccessControl to allow operations on Hive tables for DB and tables.

Fixes #59612

Signed-off-by: Amogh Margoor <amargoor@apache.org>