[Enhancement] VIEW in information_schema.tables_config (#49447) (#49776)
Why I'm doing:
In information_schema.tables_config, in TABLE_ENGINE col, if it's a view, the TABLE_ENGINE is empty.
It's better to make this column VIEW when its type is view.

What I'm doing:
Add extra logic in function generateTablesConfigResponse. If the table is OlapView or the table is ConnectorView(HIVE_VIEW or ICEBERG_VIEW), the tableConfigInfo set the table_engine col with corresponding table type.
Fixes #49447.

Signed-off-by: anyushi <434624198@qq.com>