[BugFix] Use correct parameter for pulsar_consumer.seek() (#34722)
According to the pulsar C++ API, we have 2 ways to call: consumer.seek():

    Result seek(const MessageId& messageId);
    Result seek(uint64_t timestamp);
And we intended to use the first one. But we use the wrong parameter here:

if (initial_position == InitialPosition::LATEST || initial_position == InitialPosition::EARLIEST) {
        pulsar::InitialPosition p_initial_position = initial_position == InitialPosition::LATEST
                                                             ? pulsar::InitialPosition::InitialPositionLatest
                                                             : pulsar::InitialPosition::InitialPositionEarliest;
        result = _p_consumer.seek(p_initial_position);
        ...
    }
So it call the 2nd function instead.
This PR will fix this issue.

Signed-off-by: jiutianchen <chen9t@gmail.com>