[BugFix] do not set basic auth header if credentials not provided (#54960)
Why I'm doing:
While the FE already has this logic, the BE would still send an empty Authorization: Basic header to ElasticSearch even if credentials were not provided; this causes a 403 even if the ES cluster has no AuthN/AuthZ settings. This change allows using StarRocks with ElasticSearch external catalogs even with Basic Auth disabled.

What I'm doing:
Adds a check for empty credentials (both username and password) in the ElasticSearch scan reader, in which case the Authorization: Basic ... header is not set.

Signed-off-by: Giorgio Pellero <giorgio.pellero@gmail.com>
Co-authored-by: alvin <115669851+alvin-celerdata@users.noreply.github.com>