[Enhancement] Introduce version-based vacuum control for cluster snapshot
For the vaccum of cluster snapshot, we retain all version large than latest version of the current cluster snapshot createTime
But the problem is that too many version can not be vaccumed even the files does not contained in the snapshot version.

Introduce version-based vacuum control for cluster snapshot. We get the snapshot version from cluster snapshot
and control vaccum as following:
1. Can not vacuum all retain versions.
2. Can not vacuum all files ref by the retain versions tablet meta.
3. Introduce ClusterSnapshotInfo to save the retain version information from the checkpoint image.
4. Only get the information from leader worker.

Signed-off-by: srlch <linzichao@starrocks.com>