[Enhancement] Introduce distribution_key_column_ids in TabletSchema
## What I'm doing:
This is a prerequisite for tablet spliting. We introduce distribution_key_column_ids
in TabletSchema to ensure the ability to use them for data filtering base on hash in the
future pr.

We choose column ids instead of column unique id because column ids are easier to handle
compatibility issues for the tables which create in the older version without unique id in FE.
At the same time, the column id for distribution key column will not affected by fast schema
evolution which means cids of distribution key always keep unchanged during the life cycle of
tablet.

Signed-off-by: srlch <linzichao@starrocks.com>