[Enhancement] Introduce distribution_key_column_names in TabletSchema
## What I'm doing:
This is a prerequisite for tablet spliting. We introduce distribution_key_column_names
in TabletSchema to ensure the ability to use them for data filtering base on hash in the
future pr.

We choose physical column names instead of column unique id because column names are easier to handle
compatibility issues for the tables which create in the older version without unique id in FE.

Signed-off-by: srlch <linzichao@starrocks.com>