[BugFix] Let submitted tasks without execution can be awared in starrocks::LakeServiceImpl to set a correct response and status
## Why I'm doing:
Currently, some task submitted by starrocks::LakeServiceImpl can not aware whether the task is executed or not,
which will cause the response or status to be set incorrectly.

## What I'm doing:
Introduce a Status member for AutoCleanRunnable. This st will be set to OK if it is executed. Let the cleaner
accept this internal status to set response correctly.

Signed-off-by: srlch <linzichao@starrocks.com>