[BugFix] Fix assertion failure for insert stmt planning of shuffle for primary key table
## Why I'm doing:
If we create a table satisify the following condition: 1. primary key table, 2. replication_num > 1
3. without replicate storage. In this case, if we execute a partial update insert stmt, we will get
assertion failure because we need to build the shuffle property in this case but if insert stmt is
in partial update mode, the list of ColumnRefOperator has not the same size of full schema.

## What I'm doing:
We just use outputFullSchema which generated by partial update checking instead of Table::getFullSchema()

Signed-off-by: srlch <linzichao@starrocks.com>