[Feature] Introduce Column Hash Predicate as a Record Predicate for rowset data filtering
This is a prerequisite for tablet splitting. We introduce RecordPredicate and its derived class ColumnHashModulusPredicate into RowsetMetadataPB to provide ability to filter data in all segments of the rowset using this predicate.  can be considered as a storage layer only pr
edicate to filter data and ColumnHashModulusPredicate calculate the available row for a given chunk in the following way:
1. Find columns by hash_column_names in chunk and calculate their combined CRC32 value.
2. select[i] = (hashes[i] % modulus == remainder)

Signed-off-by: srlch <linzichao@starrocks.com>