[Enhancement] Disable change file_bundling property when lake table has rollup which partitionId is not physical partitionId (#60255)
The issue(#60184) mistake the physical partitonId and logical partitionId when create LakeRollup and this issue will cause execution errors during the publish of tables with file_bundling enabled, leading to a series of problems such as query errors.
So when we upgrade from old version and try to enable file_bundling, we will check the partitionId in ShardInfo and if found mismatch partitionId, we will reject enable file_bundling.

Signed-off-by: sevev <qiangzh95@gmail.com>