[BugFix] Let submitted tasks without execution can be awared in starrocks::LakeServiceImpl to set a correct response and status
## Why I'm doing:
Currently, some task submitted by starrocks::LakeServiceImpl can not aware whether the task is executed or not,
which will cause the response or status to be set incorrectly.

## What I'm doing:
Change AutoCleanRunnable to CancellableRunnable, which has a canceller to be called when threadpool is shutdown
to handle the correct response and status. Remove the cleaner and caller should do the proper clean-up work
in run() or when the task submitted failed.

Signed-off-by: srlch <linzichao@starrocks.com>