[BugFix] Add shard group id into cluster snapshot info for GC control
## Why I'm doing:
When sr gc shard in starMgrMetaSyncer, it will remove all shard which found in starMgr but not in FE for the same shard group.
But in the previous impl, we just check the materialized index id but not the shard group id which will lead the starMgrMetaSyncer
delete the data incorrectly.

What I'm doing:
Introduce shard group id into cluster snapshot info and check it in starMgrMetaSyncer.

Signed-off-by: srlch <linzichao@starrocks.com>