[BugFix] Fix auto increment value lost when partial update in COLUMN_UPSERT_MODE for share nothing
## Why I'm doing:
In current impl, if we partial update a auto increment column in share nothing mode, we may lost the value:
1. Use COLUMN_UPSERT_MODE for partial update.
2. Key has been inserted before.
In this case, .upt files contains 0 data for the auto increment column because it expect finish the load using
the data coming from the old version, thus such 0 data in .upt files should not be used. But does use such 0
data in apply phrase get overwrite the existed auto increment column value by 0 and lost the old value.

This problem is only occur in share nothing because COLUMN_UPSERT_MODE does not supported in share data mode.

## What I'm doing:
Do not use auto increment column data in .upt file for the existed keys.

Signed-off-by: srlch <linzichao@starrocks.com>