[BugFix] Fix throw exception issue in low-cardinality optimization error in ALLOW_THROW_EXCEPTION mode
Our current process:
1. Get all dictionaries.
2. Calculate using Expr.
3. Get the mapping map.
In process 1, we get all inputs, including empty strings(""). When calculating Expr, if an exception is thrown, it will cause an error even if there are no operations that would throw an exception.

In this PR, we add a branch: if Expr throws an exception, we record the exception and return it in the final output.

Signed-off-by: stdpain <drfeng08@gmail.com>