[BugFix] Fix table with gin index use replicated_storage (#62480)
## Why I'm doing:
Table with GIN index do not support replicated_storage right now. However, even though the replicated_storage of the table has been disabled, it may still be forcibly enabled when generating the ingestion plan. Therefore, an additional check should be added at the point where replicated_storage is enabled to ensure that the table does not have a GIN index.

Signed-off-by: sevev <qiangzh95@gmail.com>