[Enhancement] Optimize trunc_to_month and trunc_to_quarter for date (#63245)
## Why I'm doing:

Given a Julian date, our current approach to obtaining `trunc_to_quarter` or `trunc_to_month` is:
1. First decode it into year, month, and day. 
2. Then, we adjust the corresponding month and day, 
3. and finally compute the new Julian date from the adjusted year, month, and day.  

However, in practice, it is unnecessary to compute a new Julian date from the adjusted values. Instead, we can directly calculate the offset between the target Julian and the current Julian based on the decoded year, month, and day.

benchmark: speedup around 50%
- trunc_to_month: https://quick-bench.com/q/fZiRaC7U1pHODHsWQuO1cvFN2qM
- trunc_to_quarter: https://quick-bench.com/q/Idw7wLB74k-5PNW2u4xcXNsoeKA

Signed-off-by: zihe.liu <ziheliu1024@gmail.com>