[BugFix] Fix dcg meta inconsistency when partial update with auto increment column in column upsert mode.
## Why I'm doing:
This bug was introduced by the pr #61341 which is fix 0 value of auto increment column in column upsert mode.
But it handle the cid incorrectly when it used to filter out auto increment column just before generate the cols file.
In pr #61341 it use txn_meta.auto_increment_partial_update_column_id() to check if the column is auto increment column.
But it is incorrect because the txn_meta.auto_increment_partial_update_column_id() means the column offset in partial segment
schema but txn_meta.partial_update_column_ids() means the column offset in full tablet schema.

## What I'm doing:
Use TabletColumn::is_auto_increment() to check if the column is auto increment column instead.

Signed-off-by: srlch <linzichao@starrocks.com>