[BugFix] Fix table compaction can not be scheduled any more after FE restart if the table compaction has to be aborted.
## Why I'm doing:
If FE restart when a compaction job is running, the corresponding txn id and table id will be recorded in the remainedActiveCompactionTxnWhenStart.
Any table in this map will not be scheduled for compaction. The problem is that the content in this map can not be removed when the compaction job is aborted
which cause the table compaction can not be scheduled any more.

## What I'm doing:
Add a method to remove the txn id from the remainedActiveCompactionTxnWhenStart when the compaction job is aborted.

Signed-off-by: srlch <linzichao@starrocks.com>