[Enhancement] Add statistic for xx_hash3_64 function (#63791)
Why I'm doing:
The function XX_HASH3_64 does not have any statistics and will be set to UNKNOWN. Even though we can not set good statistics for this function, even having trivial statistics is better then not having anything as it could be helpful for follow up functions like MOD.
For example for the expression ABS(MOD(XX_HASH3_64(a),100)) without statistics for XX_HASH3_64 the whole expression statistics will be unknown, but with trivial XX_HASH3_64 statistics starrocks can set the MIN to 0 and the MAX to 100 for the whole expression.
This could be the difference between a correct and a wrong join order.

(there are probably other cases like this one)

What I'm doing:
Adding statistic for the function XX_HASH3_64. The MIN and MAX are set to negative infinity and positive infinity. The DISTINCT COUNT is set to the number of rows.

Signed-off-by: m-selmi <m.selmi@celonis.com>