[BugFix] Consider the replica status when choosing a replica for backup.
# Why I'm doing:
In some cases, if the backend node is not available, the replica status will be DEAD.
But the replica has not any invalid version info.
Some the backup job will choose this replica for backup, but the backup job will fail.

# What I'm doing:
Add a check to the replica status when choosing a replica for backup.
If the replica status is DEAD, skip this replica.

Signed-off-by: srlch <linzichao@starrocks.com>