[BugFix] Fix full vacuum failed if tablet metadata is bundle file with InitVersion
# Why I'm doing:
In current implementation, tablet metadata with version 1 will be treated as a normal tablet metadata file even
it is a bundle file. In this case, full vacuum in BE/CN side will try to call tablet_mgr->get_tablet_metadata
to get the tablet metadata. But in this case, tablet id = 0 which cause that this function call will return NotFound error
for sure and full vacuum will failed.

What I'm doing:
When we deal with bundle tablet metadata file with version 1, we need not to call tablet_mgr->get_tablet_metadata because
it has not commit_timestamp and get_tablet_metadata basiclly is used to check this timestamp. So we need to handle this case
specially that if tablet id = 0 and version = 1 and some tabletmeta has been expired before, we can just delete it without
calling get_tablet_metadata.

Signed-off-by: srlch <linzichao@starrocks.com>