[Enhancement] Avoid OOM causing operating system freezes
StarRocks has a large binary. When the operating system runs low on memory, it attempts to free up some memory.
If the freed memory consists of executable pages that are subsequently executed shortly after, the operating system may appear to freeze.
In this patch, we use the mlock system call to lock the code segments of executable files in memory.
This can significantly reduce the chances of the operating system freezing

Signed-off-by: stdpain <drfeng08@gmail.com>