[BugFix] Consider minRetainVersion when full vacuum for partition
## What I'm doing:
In current full vacuum implementation, the vacuumable version range is [minCheckVersion, maxCheckVersion),
and maxCheckVersion is always the visible version and does not consider the minRetainVersion.
In this pr, we will consider the minRetainVersion when full vacuum for partition by modifying the maxCheckVersion
to the min of visible version and partition's minRetainVersion.

Signed-off-by: srlch <linzichao@starrocks.com>