[BugFix] Fix use-after-free when storage engine is stopped in local pk index manager.
## Why I'm doing:
For the graceful shutdown of the BE, if there are some submitted local pk index compaction tasks,
the BE exiting process will not consider the submitted tasks as finished. After ExecEnv is destoryed,
the submitted tasks still can be scheduled and access some resources that have been destroyed.

## What I'm doing:
Add the stop() method to the LocalPkIndexManager class, and call it in the StorageEngine::stop() method.

Signed-off-by: srlch <linzichao@starrocks.com>