[BugFix] Fix use-after-free when storage engine is stopped in local pk index manager.
## Why I'm doing:
For the graceful shutdown of the BE, if there are some submitted local pk index compaction tasks,
the BE exiting process will not consider the submitted tasks as finished. After ExecEnv is destoryed,
the submitted tasks still can be scheduled and access some resources that have been destroyed.

In shared nothing mode, there is a similar issue when the storage engine is stopped.

## What I'm doing:
Add the stop() method to the PersistentIndexCompactionManager class, and call it when:
1. destroy the PersistentIndexCompactionManager instance
2. stop the update manager in shared nothing mode
3. stop the storage engine in shared data mode

Signed-off-by: srlch <linzichao@starrocks.com>