[Enhancement] Add CURRENT_WAREHOUSE() information function
Add support for the CURRENT_WAREHOUSE() SQL function to return the
effective warehouse name for the current session.

This function returns the warehouse set via SET WAREHOUSE or the
default warehouse if none was explicitly selected, similar to how
CURRENT_ROLE() and CURRENT_USER() work for roles and users.

Example usage:
  mysql> SET WAREHOUSE = 'compute_pool_x';
  mysql> SELECT current_warehouse();
  +---------------------+
  | current_warehouse() |
  +---------------------+
  | compute_pool_x      |
  +---------------------+

Signed-off-by: Kevin Cai <kevin.cai@celerdata.com>