[BugFix] Clear startup lake compaction txn map on txn finalization
## Why I'm doing:
This bug was introduced in PR #63881. It move removeFromStartupActiveCompactionTransactionMap from handleCompactionFinished to
CompactionScheduler which cause a problem. When the active map was built during leader startup, such txn in this map was not
in the CompactionScheduler. Which means that these can not be accessed by CompactionScheduler and will never be removed from the map.

## What I'm doing:
Move removeFromStartupActiveCompactionTransactionMap into txn state state-machine logic and ensure that the txn is removed from the
map when it reaches a final state.

Signed-off-by: srlch <linzichao@starrocks.com>