[Tool] Python SQLAlchemy 1.4 Compatibility (#65976)
Why I'm doing:
SQLAlchemy client was built alongside SQLAlchemy 2.0, but it needs to also be compatible with version 1.4.*
This is for people/projects who have not yet been able to make the upgrade to SQLAlchemy 2.0, such as Apache Superset which quite a number of people use for visualization. Having it compatible there is a good way to improve visibility.

What I'm doing:
Tweaks to the dialect to allow it to run in SQLAlchemy 1.4. This involves removing some of the typing on methods which seems to be not fully supported by the reflection.cache decorator in 1.4. Also needed to make the starrocks_properties dictionary a hashable item so that it worked in 1.4

Signed-off-by: Pat Buxton <patrick.buxton@tartansolutions.com>
Signed-off-by: Pat Buxton <45275736+rad-pat@users.noreply.github.com>